backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "requests>=2.32.5",
    "selenium>=4.35.0",
    "webdriver-manager>=4.0.2",
]
